WIP Make fetcher a class to avoid the API issues
This CL changes Fetcher to be a function that returns value/error instead of
throwing. Internally, RealStore always uses it and any exception thrown in
such function will escalate.

There are builders in Fetcher.** which can receive different fetcher functions.

This is an alternative implementation to support fetchers that return values
and not throw exceptions.